refactor: extract SuccessButtonGroup from SuccessComponent
